Revamped select_particles() and delete_particles()
